alloydb-semantic-search/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pg8000>=1.31.0
google-cloud-alloydb-connector>=1.9.0
google-cloud-aiplatform>=1.71.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
